Interjection Classification for Improved HCI
In the realm of human-computer interaction, fostering a natural dialogue between humans and machines is paramount. A key, often overlooked, component of this dialogue is the use of interjections such as “mmm” and “hmm”. Despite their frequent use to express agreement, hesitation, or requests for information, these interjections are typically dismissed as “non-words” by Automatic Speech Recognition (ASR) engines. Addressing this gap, we introduce a novel task dedicated to interjection classification, a first in the field to our knowledge. This task is challenging due to the short duration of interjection signals and significant inter and intra-speaker variability. In this work, we present a dataset of interjection signals collected specifically for interjection classification. We employ this dataset to train and evaluate a baseline deep learning model. To enhance performance, we augment training dataset using techniques such as tempo and pitch transformation, which significantly improve classification accuracy, making models more robust.